+++ src/collective/nitf/tests/test_catalog.py
@@ -46,9 +46,9 @@ class CatalogTestCase(unittest.TestCase):
         self.assertEqual(result[0].getURL(), self.n1.absolute_url())
 
     def test_genre_indexed(self):
-        self.n1.genre = 'Current'
+        self.n1.genre = u'Current'
         self.n1.reindexObject()
-        result = self.catalog(genre='Current')
+        result = self.catalog(genre=u'Current')
         self.assertEqual(1, len(result))
         self.assertEqual(result[0].getURL(), self.n1.absolute_url())